refactor

+++ src/com/util/database/myDBHelper.java
@@ -4,6 +4,8 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 import com.model.oneMsg;
@@ -17,9 +19,15 @@ import android.database.sqlite.SQLiteOpenHelper;
 import android.os.Environment;
 import android.util.Log;
 
-/** 用来管理数据库连接的类
- * @author jumper
+
+/**
+ * 类功能描述：DBO,数据库管理</br>
  *
+ * @author 王明献
+ * @version 1.0
+ * </p>
+ * 修改时间：</br>
+ * 修改备注：</br>
  */
 public class myDBHelper extends SQLiteOpenHelper {
 	
@@ -47,6 +55,10 @@ public class myDBHelper extends SQLiteOpenHelper {
 	 * @param tableName 把tableName表清空 
 	 */
 	public void deleteRecord(String tableName ) {
+	    
+	    if (tableName == null) {
+	        return ;
+	    }
 		
 		SQLiteDatabase db = this.getWritableDatabase();
         String sql = "delete from " + tableName;
@@ -66,27 +78,29 @@ public class myDBHelper extends SQLiteOpenHelper {
 	public List<oneMsg> getChatRecord(String from,String to) {
 		
 		list.clear() ;
-		String user="";
-		String toWhom = "" ;
-		String date="";
-		String content="";
-		String dir = "" ;
+		if (from=="" || to=="")
+		    return list ;
 		
+	    String date="";
+	    String content="";
+	    String dir = "" ;
+	        
 		SQLiteDatabase db=this.getReadableDatabase();
-		String sql="select chatFrom,chatTo,chatDate,chatContent,chatDir from message_table where " +
-				"(chatFrom=? and chatTo=?) or (chatFrom=? and chatTo=?) order by chatDate";
-		Cursor cursor = db.rawQuery(sql,new String[]{from,to});
+		
+		//调出两个人的对话记录
+		String sql="select chatFrom,chatTo,chatDate,chatContent,chatDir from message_table" +
+				" where (chatTo=? and chatFrom=?) or (chatFrom=? and chatTo=?) order by chatDate";
+		Cursor cursor = db.rawQuery(sql,new String[]{from,to,from,to});
 		
 		while (cursor.moveToNext()) {
 			
-			user=cursor.getString(0);
-			toWhom = cursor.getString(1) ;
-			date=cursor.getString(2);
-			content=cursor.getString(3);
+			date = cursor.getString(2);
+			content = cursor.getString(3);
 			dir = cursor.getString(4) ;
 
 			list.add( new oneMsg(from,to, date, content, dir) ) ;
-		}
+		}	
+		db.close() ;
 		return list ;
 	}
 	
@@ -95,42 +109,18 @@ public class myDBHelper extends SQLiteOpenHelper {
 	 */
 	public void addOneMsg(oneMsg chatMsg) {
 		
-		SQLiteDatabase db=this.getWritableDatabase() ;
-		String sql1="insert into message_table(chatFrom,chatTo,chatDate,chatContent,chatDir) values(?,?,?,?,?)"; 
-		db.execSQL(sql1, new Object[]{chatMsg.getUid(),chatMsg.getToWhom(),chatMsg.getDate(),chatMsg.getMsg(),chatMsg.getDir()}); 
-	}
-	
-	/**判断表是否存在
-	 * @param tableName
-	 * @return
-	 */
-	public boolean tabbleIsExist(String tableName){
+	    if (chatMsg == null) {
+	        
+	        return  ;
+	    }
+		SQLiteDatabase db = this.getWritableDatabase() ;
 		
-        boolean result = false;
-        
-        if (tableName == null) {
-        	
-                return false;
-        }
-        
-        SQLiteDatabase db = null;
-        Cursor cursor = null;
-        
-        try {
-                db = this.getReadableDatabase();
-                String sql = "select count(*) as c from "+DataBaseName+" where type ='table' and name ='"+tableName.trim()+"' ";
-                cursor = db.rawQuery(sql, null);
-                if (cursor.moveToNext()) {
-                        int count = cursor.getInt(0);
-                        if (count>0) {
-                                result = true;
-                        }
-                }
-                
-        } catch (Exception e) {
-                
-        }               
-        return result;
+		String sql1="insert into message_table" +
+				"(chatFrom,chatTo,chatDate,chatContent,chatDir) values(?,?,?,?,?)";
+		
+		//执行SQL语句
+		db.execSQL(sql1, new Object[]{chatMsg.getUid(),chatMsg.getToWhom(),
+		        chatMsg.getDate(),chatMsg.getMsg(),chatMsg.getDir()}); 
 	}
 	
 }
